+++ accessory/characteristic/Climate.js
@@ -20,19 +20,40 @@ const CLIMATE_CONFIG = {
     tempUnit: "tempUnit", // 'Celsius' (default), 'Fahrenheit'
     minTemp: "minTemp",
     maxTemp: "maxTemp",
-    minStep: "minStep"
+    minTempStep: "minTempStep",
+    minFanSpeed: "minFanSpeed",
+    maxFanSpeed: "maxFanSpeed",
+    minFanStep: "minFanStep"
 };
 
 const DEFAULT_MIN_TEMP = -100;
 const DEFAULT_MAX_TEMP = 200;
-const DEFAULT_MIN_STEP = 0.1;
+const DEFAULT_MIN_TEMP_STEP = 0.1;
+
+const DEFAULT_MIN_FAN_SPEED = 0;
+const DEFAULT_MAX_FAN_SPEED = 100;
+const DEFAULT_MIN_FAN_STEP = 1;
+
 
 function addWaterLevelCharacteristic(service, optional) {
     addNumericSensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.WaterLevel), {item: CLIMATE_CONFIG.waterLevelItem}, optional);
 }
 
 function addRotationSpeedCharacteristic(service, optional) {
-    addNumericSensorActorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.RotationSpeed), {item: CLIMATE_CONFIG.rotationSpeedItem}, optional);
+    let thisMinSpeed = this._config[CLIMATE_CONFIG.minFanSpeed] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.minFanSpeed]) : DEFAULT_MIN_FAN_SPEED;
+    let thisMaxSpeed = this._config[CLIMATE_CONFIG.maxFanSpeed] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.maxFanSpeed]) : DEFAULT_MAX_FAN_SPEED;
+    let thisMinStep = this._config[CLIMATE_CONFIG.minFanStep] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.minFanStep]) : DEFAULT_MIN_FAN_STEP;
+
+    let rotationSpeedCharacteristic = service.getCharacteristic(this.Characteristic.RotationSpeed);
+    rotationSpeedCharacteristic.setProps({
+        minValue: thisMinSpeed,
+        maxValue: thisMaxSpeed,
+        minTempStep: thisMinStep
+    });
+
+    this._log.debug(`Applied minValue ${thisMinSpeed}, maxValue ${thisMaxSpeed} and minStep ${thisMinStep} for fan speed`);
+
+    addNumericSensorActorCharacteristic.bind(this)(service, rotationSpeedCharacteristic, {item: CLIMATE_CONFIG.rotationSpeedItem}, optional);
 }
 
 function addCurrentRelativeHumidityCharacteristic(service, optional) {
@@ -46,7 +67,7 @@ function addTargetRelativeHumidityCharacteristic(service, optional) {
 function addCurrentTemperatureCharacteristic(service, optional) {
     let thisMinTemp = this._config[CLIMATE_CONFIG.minTemp] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.minTemp]) : DEFAULT_MIN_TEMP;
     let thisMaxTemp = this._config[CLIMATE_CONFIG.maxTemp] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.maxTemp]) : DEFAULT_MAX_TEMP;
-    let thisMinStep = this._config[CLIMATE_CONFIG.minStep] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.minStep]) : DEFAULT_MIN_STEP;
+    let thisMinStep = this._config[CLIMATE_CONFIG.minTempStep] !== undefined ? parseFloat(this._config[CLIMATE_CONFIG.minTempStep]) : DEFAULT_MIN_TEMP_STEP;
 
     let transformation = this._config[CLIMATE_CONFIG.tempUnit] === "Fahrenheit" ? _convertFahrenheitToCelsius : parseFloat;
 
@@ -57,7 +78,7 @@ function addCurrentTemperatureCharacteristic(service, optional) {
         minStep: thisMinStep
     });
 
-    this._log.debug(`Applied minValue ${thisMinTemp}, maxValue ${thisMaxTemp} and minStep ${thisMinStep}`);
+    this._log.debug(`Applied minValue ${thisMinTemp}, maxValue ${thisMaxTemp} and minStep ${thisMinStep} for temp`);
 
     addNumericSensorCharacteristicWithTransformation.bind(this)(service,
         currentTemperatureCharacteristic,